+++ CHANGELOG.md
@@ -8,6 +8,13 @@ adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
 
+### Added
+
+- Add `Query` method `option` for setting parameter options as described both in
+  [CMR Search API Parameter Options](https://cmr.earthdata.nasa.gov/search/site/docs/search/api.html#parameter-options)
+  and in other sections of the CMR Search API documentation, thus supporting
+  other parameter options that are not covered in that particular section of the
+  documentation.  ([#74](https://github.com/nasa/python_cmr/issues/74))
 - Support multi-point searches ([#72](https://github.com/nasa/python_cmr/issues/72))
 
 ### Fixed